\name{star_groj0422.32}
\docType{data}
\alias{star_groj0422.32}
\title{Data: Light curve from GROJ0422+32}
\description{
Light curve for gamma ray emission of the source GROJ0422+32.
}
\usage{star_groj0422.32}
\format{A matrix of three columns, with a time series of length 729 appropriate to \code{\link{RobPer}}.}
\source{Data kindly provided by the National Aeronautics and Space Administration (NASA); National Space, Science, and Technology Center (NSSTC); Gamma-Ray Astrophysics Team (see Details).}
\details{Data obtained by the BATSE Earth Occultation Monitoring project of the NSSTC, available from \url{https://gammaray.nsstc.nasa.gov/batse/occultation/}. The experiments are described in Harmon et al. (2002) and Harmon et al. (2004).}
\references{
Harmon, B. A., Fishman, G. J., Wilson, C. A., Paciesas, W. S., Zhang, S. N., Finger, M. H., Koshut, T. M., McCollough, M. L., Robinson, C. R. and Rubin, B. C. (2002): The Burst and Transient Source Experiment Earth Occultation Technique. The Astrophysical Journal Supplement Series, 138 (1), 149-183

Harmon, B. A., Wilson, C. A., Fishman, G. J., Connaughton, V., Henze, W., Paciesas, W. S., Finger, M. H., McCollough, M. L., Sahi, M., Peterson, B., Shrader, C. R., Grindlay, J. E. and Barret, D. (2004): The Burst and Transient Source Experiment (BATSE) Earth Occultation Catalog of Low-Energy Gamma-Ray Sources. The Astrophysical Journal Supplement Series, 154 (2), 585-622
}
\note{See Vignette Section 5.2 for example.}
\keyword{datasets}